refactor: extract ScriptHandler from MCPServerV2
Move 10 script management methods into dedicated ScriptHandler.
Move _test_scripts_dir to handler; update test references.
